api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
google-cloud-firestore==2.16.1
google-auth==2.32.0
requests==2.32.3